Time: 219 ms (28.21%) | Memory: 358 MB (24.09%) - LeetSync

+++ LEETCODE/4092-finish-time-of-tasks-i/finish-time-of-tasks-i.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-long long solve(vector<long long>&dp , vector<vector<long long>>&adj, vector<int>& baseTime, int i){
+long long solve(vector<long long>&dp , vector<vector<int>>&adj, vector<int>& baseTime, int i){
     if(adj[i].size() == 0) return baseTime[i];
 
     if(dp[i] != -1) return dp[i];
@@ -19,7 +19,7 @@ long long solve(vector<long long>&dp , vector<vector<long long>>&adj, vector<int
 
 } 
     long long finishTime(int n, vector<vector<int>>& edges, vector<int>& baseTime) {
-        vector<vector<long long>>adj(n);
+        vector<vector<int>>adj(n);
         for(auto it : edges){
             adj[it[0]].push_back(it[1]);
         }
